Improved support for never versions
Mostly 3D biome stuff
