--- src/main/java/game/data/chunk/palette/JsonBlockState.java
@@ -1,11 +1,0 @@
-package game.data.chunk.palette;
-
-import java.util.HashMap;
-
-/**
- * Simple class for JSON deserialization.
- */
-public class JsonBlockState {
-    int id;
-    HashMap<String, String> properties;
-}
