--- src/main/java/game/data/chunk/palette/JsonBlockType.java
@@ -1,10 +1,0 @@
-package game.data.chunk.palette;
-
-import java.util.ArrayList;
-
-/**
- * Simple class for JSON deserialization.
- */
-public class JsonBlockType {
-    ArrayList<JsonBlockState> states;
-}
